Refatora página Huddle com novos agrupamentos e acordions
