+++ src/hooks/useHuddleList.ts
@@ -8,22 +8,8 @@ export const useHuddleList = (pacientes: Paciente[]) => {
     return pacientes.filter(paciente => {
       const dataInternacao = parseDateFromString(paciente.dataInternacao);
       if (!dataInternacao) return false;
-      return differenceInDays(new Date(), dataInternacao) > 30;
+      return differenceInDays(new Date(), dataInternacao) > 60;
     });
   }, [pacientes]);
-
-  const altasPendentes = useMemo(() => {
-    return pacientes
-      .filter(p => p.altaPendente)
-      .reduce((acc, paciente) => {
-        const tipo = paciente.altaPendente!.tipo;
-        if (!acc[tipo]) {
-          acc[tipo] = [];
-        }
-        acc[tipo].push(paciente);
-        return acc;
-      }, {} as Record<string, Paciente[]>);
-  }, [pacientes]);
-
-  return { internacoesProlongadas, altasPendentes };
+  return { internacoesProlongadas };
 };
